+++ assets/javascripts/kitten/components/layout/dashboard-layout/stories/menu.js
@@ -128,6 +128,7 @@ export const StoryMultiMenu = () => (
           Projets
         </Text>
         <ButtonWithTooltip
+          position="left"
           buttonProps={{
             fit: 'icon',
             size: 'nano',